+++ g2gmobile/package.json
@@ -11,17 +11,17 @@
     "babel-preset-react-native-stage-0": "^1.0.1",
     "expo": "^17.0.0",
     "immutability-helper": "^2.2.2",
-    "mobx": "^3.1.10",
+    "mobx": "^3.1.16",
     "mobx-logger": "^0.6.0",
     "mobx-react": "^4.1.8",
-    "native-base": "^2.1.4",
+    "native-base": "^2.1.5",
     "react": "16.0.0-alpha.6",
     "react-native": "https://github.com/exponent/react-native/archive/sdk-17.0.0.tar.gz",
-    "react-native-simple-store": "^1.1.0",
-    "react-redux": "^5.0.3",
+    "react-native-simple-store": "^1.3.0",
+    "react-redux": "^5.0.5",
     "redux": "^3.6.0",
-    "redux-actions": "^2.0.1",
-    "redux-logger": "^3.0.0",
+    "redux-actions": "^2.0.3",
+    "redux-logger": "^3.0.6",
     "redux-promise": "^0.5.3"
   }
 }
